+++ src/readers.rs
@@ -197,7 +197,9 @@ impl crate::FileFormat {
         };
         for line in reader.take(READ_LIMIT).lines().take(LINE_LIMIT) {
             let buffer = line?.to_lowercase();
-            if buffer.contains("<collada") {
+            if buffer.contains("<amf") {
+                return Ok(Self::AdditiveManufacturingFormat);
+            } else if buffer.contains("<collada") {
                 return Ok(Self::DigitalAssetExchange);
             } else if buffer.contains("<mxfile") {
                 return Ok(Self::Drawio);